Indent prm files in examples folder and documentation (#671)
Description of the problem
The indentation of all the parameter files in the examples folder was not uniform.
The extracts of the parameter file in each example in the documentation was not properly indented.

Description of the solution
The prmindent file located in the utilities folder was used to indent all parameter files in the examples folder.
Tabs were removed and proper indentation was applied to the parameter file sections of each example in the documentation.

Documentation
A warning was added in the "contributing" tab of the documentation explaining how to properly indent sections of a parameter file when writing a new example.

Comments
While fixing the indentation of the examples I noticed several differences in some examples between the parameter files explained in the documentation and the actual parameter files in the examples folder. This should be fixed to avoid confusion.