Optimize particle-wall fine search and emplace (#673)
Description of the problem
Tuples generated when creating the particle_wall_fine search where created then inserted. Same was being done for the contract info
Particle wall fine search did not use search radius

Description of the solution
Tuples are now emplaced directly in the container for everything that was possible related to particle-wall contact. This actually has a much larger impact on performance than I expect (like... close to 5-6% on some cases)
I tried using the serach radius for the particle-wall contact, but it turns out this breaks cases when the walls are moving (and this makes sense, static particle + moving wall won't be registered as a contact. Consequently, I don't think this change is worth it for now.

How Has This Been Tested?
All tests should pass. If so, everything is good.

Documentation
No changes

Comments
Can't seem to give up on keeping on optimizing this thing.