Granular dam break example (#677)
Description of the problem
There was only a single example with DEM that had moving geometries

Description of the solution
Add a granular dam break example. It's not the richest example ever, but it is quite simple and straightforward and I appreciate that of it. There is also a small comparison with experimental results and the agreement is quite good even though I had to scale up the particles significantly. The example is also very fast to run (8 minutes on my computer)

Documentation
Example is fully documented

Future changes
Could be interesting to try with more than one height. Will do this as an addendum in following days

Comments
I have a few examples like this in the oven, will make PRs with them soon.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Former-commit-id: 2285af7f4214c355a0fbf46368e6eda664033494