Add thermal post-processing to nitsche IB (#678)
Description of the problem
The nitsche beta heat coefficient for heat transfer was scaled with the mesh size. This enabled good dirichlet boundary conditions but made it difficult to carry industrial analysis with it.
There was no way to post-process the heat extracted by the Nitsche IB

Description of the solution
The beta heat coefficient is now mesh independent. This parameter now has a physical meaning, it is the heat extracted W/m^2 (in 2D) or W/m^3 in 3D
The nitsche heat flux is now post-processed when other heat fluxes are calculated.

How Has This Been Tested?
A new test has been added to validate this approach

Documentation
Documentation has been modified accordingly

Co-authored-by: Jeanne Joachim <72091781+jeannejoachim@users.noreply.github.com>
Co-authored-by: jeannejoachim <jeanne.joachim@gmail.com>