Dem optimize particle-particle forces by templating rolling friction model (#676)
Description of the problem
Rolling friction was still passed as a parameter and the model was checked using an if condition for all partice ij pair

Description of the solution
The type of rolling friction model has been templated. The if is now a if constexpr so this prevents billion of ifs checks

How Has This Been Tested?
All tests pass

Comments
I tried a ton of optimizations and I have not found anything else where I could increase the speed of the simulations. I think we are reaching the limit...

Former-commit-id: f57a4216971932c350b99443fe6b8d8df9389cbf