Add vof phase factor in heat transfer laser source and radiation term at the free surface (#683)
Description of the problem
For additive manufacturing simulations the both the heat transfer and VOF solvers, there were two required changes:

one issue to be addressed: the metal AND the air were heated by the laser instead of only the metal.
one feature to be added, that is the radiation cooling at the metal/air interface, which is a significant cooling mechanism in additive manufacturing simulations.
Both changes were implemented in the heat transfer solver:

For the air "permeability" to the laser, the class HeatTransferAssemblerLaserVOF was implemented. The laser source term is computed as in the previously implemented class HeatTransferAssemblerLaser, however, it is multiply by the phase fraction (computed in the VOF solver) to activate the laser source term in the metal (phase fraction = 1) and deactivate it in the air (phase fraction = 0).
For the radiation cooling, the class HeatTransferAssemblerFreeSurfaceRadiationVOF was implemented. Also, the parameters struct Laser_FreeSurfaceRadiation was implemented to described the required parameters for the radiation term and was added as an attribute of the parameters class Laser. The radiation term is applied only in the cells at the air/metal interface. It is treated as a source term instead of a boundary term using the phase fraction gradient as the interface indicator.
Both features are enabled in the HeatTransfer class.
Description of the solution
The air is now "permeable" to the laser heating source, i.e., it is not heated by the laser.
The radiation term was added (following the Stefan-Boltzmann law of radiation) and verified with a steady-state analytical solution.
How Has This Been Tested?
[applications_tests/gls_navier_stokes_2d/heat_transfer_vof_radiation.prm and .out] An application test with an analytical solution was added for the radiation term. The computational domain is a square of an unit side length: the bottom half is fluid 1 (metal) and the top half is fluid 0 (air). The radiation term is applied at y=0.5.
all the other applications tests passed.
Documentation
[doc/source/parameters/cfd/laser_heat_source.rst ] The added parameters for the radiation term were documented.
Future changes
An example will be added.