Refactor cells iteration for mobility checks
Before this modification, there was 5 loops over all cells (loops for granular temperature & 4 mobility at nodes).
Now, 1 loop is done in the granular temperature calculation and a set of local and ghost cells is created. During mobility checks at nodes, there's an iteration over the set and cells with known mobility are removed from the set. This prevent next checks to check the mobility of those cells