Improve rising bubble (#696)
Description of the problem
The results of the rising bubble simulation example were of very poor quality
The boundary conditions on the example were not correctly posed
There was an error in the weak form and in the implementation of the surface tension
There was a weird element in the calculation of the curvature.

Description of the solution
Fixed the boundary condition
Added new post-processing capabilities to be able to make an adequate comparison with the literature.
Updated animation and graphs
Updated documentation
Updated example to also use the new filtering methods. This also helps the result a lot.
Fixed weak form
The output of some tests has been changed because of this. I am confident in the benchmark results now however.

How Has This Been Tested?
Example has been re-run with this post-processing capability
Documentation
Added documentation of the new feature
cleaned the VOF documentation. There were so many tips and tricks and pop-up bar that it was very difficult to see what was going on
Updated the example
Future changes
make all VOF example use filtering.

Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>

Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>