Add writting and array modifying features to Lethe_pyvista_tools (#686)
Description of the problem
We wanted to create a post-processing tool to create or modify an array according to a given condition and resulting values.
We also wanted to be able to create new vtu and pvd files so that it was possible to see the modifications on Paraview

Description of the solution
Add modifier to post-processing module: The modifier takes a string with a condition for the position ("x", "y", "z"), velocity ("u", "v", "w"), time ("t"), or any other array and modifies (or creates) an array according to the condition;
Another argument takes the resulting value, modifying function as a string, or a numpy array, and uses it to modify or create the array.
A writing function was also implemented.

How Has This Been Tested?
 Liquid fluidized bed
 Hopper
 small rotating drum

Documentation
 Add example file.
 Add application example

Future changes
Examples will be added to the documentation in the near future.

Co-authored-by: Audrey Collard-Daigneault <71884806+acdaigneault@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>

Double-check indentation