New cad sharp ib (#691)
Description of the problem
-This PR enables the sharp IB solver to model particles from a step, iges, or stl file using OpenCascade.
-To make this possible, we made significant modifications to the structure of the level-set evaluation. These modifications aim at reducing the number of evaluations of the level set.
-This PR also introduces the extra verbose output option in the sharp ib solver. This option enables the output of a new output field in the .vtu file. Currently, this new output field indicates which particle cuts each cell.
-The modification in this PR requires installing dealII and Lethe with OpenCascade enabled.
-This PR also fixed some issues related to the force evaluation of nonspherical shapes.
-This PR also fixed some issues with the interaction of non-newtonian fluid and weak boundary conditions.

Description of the solution
-A new shape called opencascade has been added to the library of shapes the sharp IB solver can model.
-A cache system has been added to reduce the number of evaluations required of the level set.
-We added a viscosity scale variable to the assembler. This variable is used in the weak boundary condition imposition and the Laplace equation assembled inside the particle.
-The force evaluation now uses a manifold given by shape to evaluate the force at the surface combined with a new evaluation.
-The normal used in calculating the force has been redefined since it was based on a spherical manifold. This corrects some errors in the force evaluation obtained for a nonspherical object.

How Has This Been Tested?
Multiple tests have been modified to fit the new and improved force evaluation
A new test: static_stokes_step, has been added. This test is identical to the static_stokes test but is performed using a sphere defined by a step.

Documentation
The documentation has been added to an upcoming PR named sharp_doc.
Still needing to be added in the documentation is how to install Lethe with OpenCascade enabled. Idem for using it on compute Canada.

Future changes
This opens the doors to a large array of possibilities. In the mid-term, we would like to add the capability to simulate contact between spherical particles and nonspherical objects.
The addition of the resolution of the coupled motion of these objects would be relatively simple. It would only require the
addition of Eulers angle integration.

Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: oguevremont <guevremont.o@gmail.com>
Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>
Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Former-commit-id: 983cec514a3f52ee6e76388310136d0d2d8adce1