Automatic case generator and launcher (#716)
Description of the problem
When doing mesh sensitivity, convergence analysis, force/torque analysis, we usually want to launch multiple cases of the same problem, and change some parameters in the .prm file (such as the kinematic viscosity or the characteristic length of the mesh). To do so, we need an easy script, that allows us to generate automatically the cases, and launch them (also automatically) locally or on the cluster.

Description of the solution
Creation of python scripts that generates automatically folders, containing a different .prm file describing the case to simulate.
Creation of python scripts that launch automatically Lethe, locally and on a cluster.
Creation of an example to better explain the scripts: cases of the 2D flow around a cylinder with variation of the inlet velocity.
Minor changes in the 3D ribbon mixer (SRF) example, so that the scripts matches the style of the new template.

How Has This Been Tested?
Locally and on Beluga. Works perfectly fine.

Documentation
[ automatic_launch.rst]: Documentation about the scripts using the 2D flow around a cylinder example.


Former-commit-id: 13f2309619221a71eeb6ab85ff7f74aba055ab93