Add description of linear solvers documentation (#721)
Description of the problem
There was an empty linear solvers page in the theory guide of the documentation.
Description of the solution
All the solvers were briefly described and useful references were added.
Documentation
 Modified file: doc/source/theory/fluid_dynamics/linear_solvers.rst
Comments
I tried to be concise and to explain the general characteristics of the solvers. I can write more details but I do not think that this is the point of the documentation, instead, I decided to point to relevant references.

Former-commit-id: a65ab247931c96cf81be6331851c319f4a13aa7c