Optimize floating mesh mapping into background triangulation (#722)
Description of the problem
The current DEM solver map every triangle of the floating meshes inside of the background triangulation for the particle-floating mesh contact detection. Although robust, this is very slow and not practical.

Description of the solution
I have added a new mechanism where the triangles are mapped in the background triangulation using the distance between the center of the cell and the triangles

How Has This Been Tested?
I tested it on all the examples and it works well
A unit test has been added for the mapping
An application test has been added with a slowly moving wall. I used this to debug also
I took the liberty of improving the granular mixer case. It was far too slow.
Documentation
[] I need to finalize the granular mixer documentation with the updates i did to the case.

Future changes
I will keep on optimizing this in the future, but this is a first work.

Comments
This already gives a considerable speed-up. The dambreak case on my machine went from 8.5 minutes to 5.5 minutes. The mixer case is even more intense.

Co-authored-by: Audrey Collard-Daigneault <71884806+acdaigneault@users.noreply.github.com>

Former-commit-id: e2ec6fa3c131ecae5a5f1162dd30682c38d07115