Improve assembly flow (#730)
Description of the problem
The current matrix assembly mechanism has some sub-optimal operations, especially with respect to second-order tensor. This PR aims at addressing this issue

Description of the solution
Depending on the values of the component, some matrix assembly steps can be skipped. This reduces the computational complexity signifiantly. This PR improves performance for Q1-Q1 matrix assembly by 10-12% and Q2-Q2 by 30%

How Has This Been Tested?
All current test pass without a hiccup

Comments
I tried the same principle with the RHS but it did not improve performance at all (decreased it actually)
This process should be done for the other Navier-Stokes assembler in future PRs