Add disabling contacts of particles with negligible motion (#720)
Description of the problem
Some DEM/CFD-DEM type of simulation has areas in their domain where particles are almost motionless, this results of irrelevant computation of contact list, contact forces and velocity integration where all particles are constantly vibrating in place
The potential of the calculation of the granular temperature was not fully exploited

Description of the solution
A feature that can associate status to cells regards the mobility of their particles has been implemented. When disabling particles' mobility is enabled, this feature is called after the "sort_particles_into_subdomains_and_cells"
The general idea behind the algorithm :
It uses the granular temperature to determinate if particles in a cell are mobile enough that the contact forces are worth computation. A cell having a granular temperature under a value (default is 1e-4) will have an inactive status which makes them rejected in the broad search step. This results in no contact forces or velocity integration in future steps for those particules, which make less computation cost for the simulation.
Cells may be flagged as that called mobility status : mobile (everything is calculated like if feature is not enabled), active (motionless particles next to mobile particles, see above) or inactive (particles are not in a neighbor’s candidate list of cells around)
There are some edge cases that need some attention. In these cases, particles in cells can't be deactivated :
The solid fraction of the cell is under a value (default = 40%) : particles in the cell may have other forces that can move them and it is supposed that with this fraction, there is not enough particles around to hold them in their position.
Cells having empty cell neighbors : it means that particles are next to a floating wall/mesh and they may have a change of contact forces from it if the wall disappears or is moving.
The cell has cell neighbors which is flagged as mobile from the 3 criteria mentioned above (granular temperature, solid fractions or floating walls) : this is because motion is badly propagated to particles to the cell around without this additional "layer" of mobile cells.
The cell has cell neighbors which is flagged as mobile from the previous criterion (additional layer of mobile cells) : this is again to allow the propagation of motion, but only the particles in contact with particles from the mobile cells are considered for the contact force calculation but their position is not computed at the integration step. Those cells are flagged as active cells.
- It works with the assignment and verification of the mobility status at nodes to check the status of the neighboring cells. (See in the comments section.)


How Has This Been Tested?
Computational gain is significant for the hopper case (2x faster) where some of the particles are motionless during loading and when there's an accumulation of particles at the bottom.
Mass flow rate results for the hopper have no significant difference (<1%)
An application test was developed to evaluate each case of status for dem_3d (mobility_status.output)
An application test was developed for the load balancing with weight according to the mobility status (load_balancing_mobility_status.output)

Documentation
No full documentation for now : this feature will be improved in the future and tested more rigorously (especially for CFD-DEM) before.

Future changes
Modification regards testing results interpretation in CFD-DEM
Add a full example (maybe pneumatic conveying with CFD-DEM)
Comments
A trivial version with checks to the neighbor cells state (and not the node status) was first made, but this results in issues using multiprocessing. Since there are at least 3 stages of information of cells (e.g., 1st cell from granular temperature value, 2nd cell from the additional layer of mobile cells and the 3rd from the active cells layer), having the data from the locally owned and ghost cells is not enough (only 2 stages of information is available in some case for a processor), we need to exchange information. To exchange this information, having information at nodes allows using the LinearAlgebra::distributed::Vector where data is exchanged with the easy call of update_ghost_values(). But this is harder to understand.
Functions in broad search classes (p-p, p-w, ...) are copy pasted and slightly modify to evaluate the cell mobility status. This doesn't look good, in my opinion, but this allows preventing lots of "if calls" for no reason when the feature is not enabled.

Co-authored-by: Victor Oliveira Ferreira <65647639+voferreira@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>