Improve CFD-DEM workflow (#732)
Description of the problem
There were multiple inconsistencies between the CFD-DEM solver, the DEM solver and the Navier-Stokes solvers.
There were some parameters which were duplicated

Description of the solution
I tried to refactor the workflow of the simulation to streamline it. I deleted duplicate parameters
I also took the time to improve the examples, they launch much quicker than they used to because they are better parametrized
The CFD-DEM now also outputs the same statistical information as the DEM
It is now clear in the CFD-DEM when the VANS equations are being solved and when the Volume-Averaged fluid dynamics is being solved
How Has This Been Tested?
All tests have been updated but there results have not changed


Comments
There are very very few test for the CFD-DEM. Notably, none include load balancing. We need to generate more application tests for the solver to ensure it remains robust.

Co-authored-by: Victor Oliveira Ferreira <65647639+voferreira@users.noreply.github.com>
Co-authored-by: Victor Oliveira Ferreira <65647639+voferreira@users.noreply.github.com>
Co-authored-by: Victor Oliveira Ferreira <65647639+voferreira@users.noreply.github.com>
Co-authored-by: Toni El Geitani <55159171+tonielgeitani@users.noreply.github.com>