Update simulation control verbosity and tests (#733)
Description of the problem
The delimiter for the simulation control output to the terminal was always a bit off sometimes. Less * than needed, more than needed, and so on and so forth. @voferreira did not like that :)

Description of the solution
Use the announce string function to print it instead so that the width is always correct.

How Has This Been Tested?
Does not change anything. However, it modifies the output of ALL tests. Consequently, I decided to do this on this small branch.

Future changes
I think uniformity is good. So trying to normalize our outputs is a good idea

Former-commit-id: 9e080b1ac694b1ab9ff2182adc56446dfe417fab