Implement viscous dissipative torque in the Unresolved CFD-DEM (#738)
Description of the problem
No viscous torque model was available in the unresolved CFD-DEM, so the contact with the fluid did not dissipate particle rotational momentum.
This leads to very unphysical results in terms of particles' rotational velocity

Description of the solution
Implemented the simple model in Bird's Transport Phenomena book to calculate the torque on a particle.
To do so, fem_torque was added to dem_properties and applied back in the particles when viscous torque is applied.

How Has This Been Tested?
It works to dissipate the rotational velocity on a sedimenting particle with a given initial omega.
Documentation
The viscous torque parameter was added to the documentation.

Future changes
Soon, a more complete model, but with a similar structure, will take its place.

Comments
Thanks @acdaigneault and @Luckabarbeau for helping me debug the IB incompatibility with the new dem properties.
Adding the 3 DEM properties did not increase the total wallclock time of the unresolved CFD-DEM case sedimentation of 10000 particles. I compared the average among 100 time-steps. The tests were done in 6 cores on my machine (Intel Core i7 - 10750H 2.60 GHz, 32 Gb RAM)
---------

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Former-commit-id: dd2e2aaaa968d304a7e24d8dd69d0285a5aabe0f