Modify the dynamic flow control and add it to CFD-DEM (#746)
Description of the problem
The dynamic flow control wasn't working for CFD-DEM since it didn't take the void fraction into account
It wasn't possible to adjust the flow convergence rate or the stability resulting of large flow rate/average velocity oscillations due to large difference between beta force and the new beta force values
There was some issue with the void fraction periodicity with the GLS VANS solver

Description of the solution
Implementation of the flow control in the coupling CFD-DEM solver
Modify the type of the target value: flow rate has been changed to average velocity. It is a better concept for CFD-DEM since we can't evaluate the flow rate at a boundary, we need to evaluate the velocities and the void fraction for the whole domain. In other to keep the flow controller model the same for CFD and CFD-DEM (volumetric flow rate vs average velocity), I choose to also change the type value evaluation for CFD. Also, the average velocity value is always + if the flow goes in the domain, it the opposite for the volumetric flow rate because the normal vector at inlet was considered.
Add a relaxation coefficient (alpha) that may help control convergence or stability
Add a parameter allowing beta force for particles: it is calculated as beta_p = beta_f * rho_f / rho_p [m/s²] and the force is apply to all particles
Resolve the issue for the void fraction periodicity with make_periodicity_constraints
Add a beta threshold so beta is not changed when the new value is in the range (default is 0.0). May be useful for slight variation.
Some refactoring of the dynamic flow control has been done :
calculate_average_velocity() (previously calculate_flow_rate()) and calculate_beta() are now executed at the same place in the code, at the end of the iteration. Before this modification, the flow rate n was calculated at the end of a step n and the beta n+1 was calculated at the beginning of the step n+1. Now, both flow rate n and beta n+1 are calculated at the end of the time step n. You may see that the printed average velocity (flow rate) of step n+1 is actually the value of n.
How Has This Been Tested?
All tests for flow control has been reviewed. They were modified and tested with the new parameters.
For the application tests, the target value is now average velocity = volumetric flow rate / area of the old test version so the calculated beta values are the same.
The result for the unit tests has been calculated in a spreadsheet and the printed results are the same.
The parameters of the Periodic Hills example has been modified and tested

Documentation
The documentation is updated with the new parameters (average velocity target, alpha, enable beta particle, beta threshold)
The term "volumetric flow rate" is now "average velocity"
Add the equation of the main controller in doc

Comments
The flow control for CFD-DEM has not been fully tested (not with qcm void fraction calculation method). This branch is a part of another branch that has other fix (qcm is not currently working on master in periodic).
I'm planning to do a application test with the flow control with qcm in periodic on the branch that fix the qcm.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: oguevremont <guevremont.o@gmail.com>