naca0012-example (#741)
Description of the problem
2D NACA0012 low Reynolds example is finally out
There were examples using transfinite meshes, but none that explained how it worked.
A full example was added along with its
Description of the solution
I made a .geo file with all the steps explained. In there you can learn how to generate curves given by a function, how to use transfinite meshing, and how to mix script/GUI to create a mesh
How Has This Been Tested?
The results obtained were compared to the literature and were quite satisfying
Documentation
If applicable, list the documentation files changed and their changes

[ lethe/lethe/doc/source/examples/incompressible-flow/incompressible-flow.rst] modified to insert the new example

[lethe/lethe/doc/source/examples/incompressible-flow/2d-naca0012-low-reynolds] added

Future changes
When I manage to find a good technique, I'll try to complete the example by computing the abscissa of the reattachment point of the flow in the upper naca.