Full optimized GLS stabilization with LSIC term (#749)
Description of the problem
The current GLS stabilization was a bit innefficient
It was also missing the LSIC term

Description of the solution
Used the same optimized mechanism as the PSPG-SUPG assembly
Added the LSIC term

How Has This Been Tested?
Tested on the unit test (whose results have chanegd slightly). Using the examples (Taylor-Couette, etc.) I have also tested the implementation at very high-order. Everything works well.

Documentation
This will need to be added to the theory guide eventually (over the course of the Summer).

Comments
Will need to check if this actually does something different on canonical turbulence benchmark

Co-authored-by: Laura Prieto Saavedra <lauraprietosaavedra@gmail.com>