Matrix-free advection-diffusion prototype (#752)
Description of the problem
This adds a matrix-based and a matrix-free advection-diffusion prototype with stabilization for different test cases.

Comments
This is the second step towards matrix-free Lethe. The boundary conditions are a bit messy since there are different cases, however, we decided to leave it like this as it is only a prototype.