VOF - Implement filtered phase in surface tension force calculation (#751)
Description of the problem
After implementing the filtered phase for the calculation of physical properties in the VOF solver, a consistency issue was noticed in the calculation of the surface tension force (STF) as the filtered phase was not considered.

Description of the solution
In the STF term of the continuum equation, the phase fraction gradient is used to calculate the normal (in the curvature expression) and to estimate the Dirac delta see (static bubble example). The phase fraction gradient is now calculated with the filtered phase. Modifications were made to both the NavierStokesScratchData class and HeatTransferScratchData class.
Rising bubble example's "tanh" filter's beta parameter value was increased to 20 (previously 10).
Replace "filtered" term by "projected" when calculating the phase fraction gradient's projection to avoid confusion.

How Has This Been Tested?
 gls_vof_tanh_filter.output was updated
 Rising bubble and static bubble examples (comparing results)
 new test (gls_vof_static_droplet_surface_tension_force_with_tanh_filter) combining STF and the "tanh" filter

Documentation
A few minor corrections (typos) were made in the rising bubble and the static bubble example files.

Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>