Add high-order output for the boundaries (#756)
Description of the problem
High-order boundary mesh was not correctly exported when "output boundaries = true". This was an annoyance in some simulations as noted by @acdaigneault 
Description of the solution
Fixed it and this closes Boundary mesh generated is not high-order #727

How Has This Been Tested?
Tested it on a dummy Poiseuille case. The results are convincing. Before and after pictures are below.