VOF boundary conditions (#768)
Description of the problem
When solving different cases, boundary conditions were only applied to N-S equations. Therefore, when applying boundary conditions such as a periodic boundary condition, the right fluid was not necessarily transported on the opposite side of the domain. In the same way, until now, it was not possible to specify which fluid would enter through an inlet or exit through an outlet.

Description of the solution
Periodic boundary condition for the phase fraction was implemented implicitly in the code when a periodic boundary condition is specified in the boundary conditions subsection of the parameter file.
Dirichlet boundary condition for the phase fraction was implemented explicitly with the new dirichlet type in the boundary conditions VOF subsection of the parameter file.
How Has This Been Tested?
Two tests were added to test the two types of boundary conditions.

 applications_tests/gls_navier_stokes_2d/gls_vof_periodic_boundary_condition
 applications_tests/gls_navier_stokes_2d/gls_vof_dirichlet_boundary_condition

Documentation
In the multiphysics boundary conditions' documentation, the new boundary conditions were added.

 doc/source/parameters/cfd/boundary_conditions_multiphysics.rst
Comments
A gravity wave example folder was also added to multiphysics examples. This example will be completed after further improvements of the VOF model.

Former-commit-id: 90e63871861f675e47b5a1d572620e2d1d7b789f