Fix QCM for particles with PBC (#763)
* Preliminary set up for pbc for qcm

Add variable, flags and mapping of periodic vertices

* Simplify handling of the reference sphere in qcm

* Add periodicity to qcm calcuDescription of the problem
The current implementation of the quadrature-centered method has not working with particles at periodic boundaries. The QCM reference sphere was not able to take into account the particles in the cells at the periodic boundary.
Description of the solution
Mapping of cells having coinciding vertices (periodic vertices) was added to handle quadrature points and particles in periodic cells.
Some refactor was done for the radius of the reference sphere in other to minimize the calculation when the value is user defined – independent of the current cell size.
How Has This Been Tested?
The void fraction of fluid around a particle crossing a PCB was compared to the one around a particle in the domain. The particles are separated by half the length of the domain for a better comparison (meaning both halves of the domain should gives same results). The animation shows the results before and after modifications. Before the modifications, the void fraction values at quadrature points at PBC was not correct. Also when the particles crossed the PBC, the particles wasn't even in a reference sphere, so gives no volumetric contribution.
vf

A application test has been added for CFD-DEM coupling 3D based on the previous case (animation), but capturing only a few time steps where the particle crossed the PBC. This test uses dynamic flow control as promised here : Modify the dynamic flow control and add it to CFD-DEM #746
Documentation
The doc is not affected by this modification
The default value of the bound void fraction was fixed in doc according to the default value in the parameter parsing
Reference of Toni's paper about QCM was added in the QCM theory
Comments
I'm curious if there a better way to handle the cell neighbors instead of using the vertices_cell_mapping and find_cells_around_cell. I feel we could map the neighbor cells directly as done in DEM solver, but maybe this usage was done of a specific reason.
I tried to change vertices to cells map to an ordered map, but since it was also used in resolved CFD-DEM, I didn't want to go there...