Add box refinement (#773)
Description of the problem
Using box refinement was impossible for these solvers.
Description of the solution
The usage of the appropriate function was added.
How Has This Been Tested?
All tests still pass.
Documentation
No applicable. The documentation already said that it was possible to use box refinement in all solvers.

Former-commit-id: b6d3461b05c3b8064fce65b3ce57d17c48c33358